fix(images): Move emitESMImage util to another file, so we don't import it by accident in contexts we shouldn.t
