+++ packages/astro/src/assets/vite-plugin-assets.ts
@@ -9,9 +9,10 @@ import type { AstroPluginOptions, ImageTransform } from '../@types/astro';
 import { error } from '../core/logger/core.js';
 import { joinPaths, prependForwardSlash } from '../core/path.js';
 import { VIRTUAL_MODULE_ID, VIRTUAL_SERVICE_ID } from './consts.js';
-import { emitESMImage, isESMImportedImage } from './internal.js';
+import { isESMImportedImage } from './internal.js';
 import { isLocalService } from './services/service.js';
 import { copyWasmFiles } from './services/vendor/squoosh/copy-wasm.js';
+import { emitESMImage } from './utils/emitAsset.js';
 import { imageMetadata } from './utils/metadata.js';
 import { getOrigQueryParams } from './utils/queryParams.js';
 import { propsToFilename } from './utils/transformToPath.js';
